requirements.txt
hydra-core>=1.3.0
omegaconf>=2.3.0
google-cloud-aiplatform>=1.38.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
checkov>=3.2.0
detect-secrets>=1.4.0
terraform-compliance>=1.3.0
